Creating previous week view

+++ tasks/views.py
@@ -1,8 +1,4 @@
 from datetime import date
-from django_cte import With
-from django.db.models import Q
-from django.urls import reverse
-from django.utils import timezone
 from django.db.models import Count
 from dateutil.relativedelta import relativedelta
 from django.shortcuts import render, redirect, get_object_or_404
@@ -11,23 +7,26 @@ from tasks.forms import TaskForm
 from tasks.models import Task, TaskCompletions
 from utils.get_date import get_last_day_of_week, get_first_day_of_week
 
-# Create your views here.
+
 def index(request):
     first_day_of_week = get_first_day_of_week(date.today())
     last_day_of_week = get_last_day_of_week(date.today())
 
-    _pending_tasks = pending_tasks() 
-    summary = summary_of_week()
+    _pending_tasks = pending_tasks(first_day_of_week, last_day_of_week) 
+    summary = summary_of_week(first_day_of_week, last_day_of_week)
 
     count_pending, count_completed = 0, 0
-    for task in pending_tasks():
+    for task in _pending_tasks:
         count_pending += task['desired_weekly_frequency']
         count_completed += task['completion_count']
     
-    count_completed_percent = (count_completed * 100) // count_pending
+    try:
+        count_completed_percent = (count_completed * 100) // count_pending
+    except:
+        count_completed_percent = 0
 
     if request.method == 'POST':
-        try:
+        if request.POST.get('task'):
             task_sent = request.POST.get('task')
             task = get_object_or_404(Task, id=task_sent)
             task_completion = TaskCompletions(
@@ -35,12 +34,14 @@ def index(request):
             )
             task_completion.save()
             return redirect('tasks:index')
-        except:
+        elif request.POST.get('id_completion'):
             id_completion = int(request.POST.get('id_completion'))
             TaskCompletions.objects.filter(id=id_completion).delete()
             return redirect('tasks:index')
+        else:
+            print('entrou aqui poha')
+            return redirect('tasks:previous_week')
 
-    # print(summary)
     return render(
         request,
         'tasks/index.html',
@@ -84,10 +85,7 @@ def create(request):
     )
 
 
-def pending_tasks():
-    first_day = get_first_day_of_week(date.today())
-    last_day = get_last_day_of_week(date.today())
-
+def pending_tasks(first_day, last_day):
     tasks_created_up_to_week = Task.objects\
         .values("id", "title", "desired_weekly_frequency")\
         .filter(created_at__range=(first_day, last_day))
@@ -97,12 +95,6 @@ def pending_tasks():
         .annotate(completion_count=Count("task_id"))\
         .filter(created_at__range=(first_day, last_day))
     
-    # ADICIONAR TUDO AGORA NO MESMO DICIONARIO, VALIDANDO DE O ID DA TESK É IGUAL O TASK_ID DA COMPLETIONS
-    # SE DER CERTO AJEITAR CÓDIGO QUE ESTÁ UMA MERDA
-    # print()
-    # print(tasks_created_up_to_week)
-    # print(task_completion_count)
-    # print()
     for task in tasks_created_up_to_week:
         for completion in task_completion_count:
             if task['id'] == completion['task_id']:
@@ -112,21 +104,10 @@ def pending_tasks():
         if not 'completion_count' in task:
             task['completion_count'] = 0
     
-    # for task in tasks_created_up_to_week:
-    #     print(task)
-
     return tasks_created_up_to_week
 
-    # return render(
-    #     request,
-    #     'tasks/index.html'    
-    # )
 
-
-def summary_of_week():
-    first_day = get_first_day_of_week(date.today())
-    last_day = get_last_day_of_week(date.today())
-
+def summary_of_week(first_day, last_day):
     tasks_created_up_to_week = Task.objects\
         .values("id", "title", "desired_weekly_frequency")\
         .filter(created_at__range=(first_day, last_day))
@@ -154,13 +135,33 @@ def summary_of_week():
 
     return summary
 
-# TESTS 
-# from itertools import chain
-# print(list(chain(tasks_created_up_to_week, task_completion_count)))
-# pending_tasks = Task.objects.values('title', 'desired_weekly_frequency')\
-#     .with_cte(tasks_created_up_to_week)\
-    # .select_related(task_completion_count+)
+
+def previous_week(request):
+    first_day_of_week = get_first_day_of_week(date.today() - relativedelta(days=7))
+    last_day_of_week = get_last_day_of_week(first_day_of_week)
+
+    _pending_tasks = pending_tasks(first_day_of_week, last_day_of_week) 
+    summary = summary_of_week(first_day_of_week, last_day_of_week)
+
+    count_pending, count_completed = 0, 0
+    for task in _pending_tasks:
+        count_pending += task['desired_weekly_frequency']
+        count_completed += task['completion_count']
     
-# test = tasks_created_up_to_week.union(task_completion_count)
-# pending_tasks_2 = TaskCompletions.objects.values('task_id', 'completionCount')\
-#     .with_cte(task_completion_count)
+    count_completed_percent = (count_completed * 100) // count_pending
+
+    if request.method == 'POST':
+        return redirect('tasks:index')
+
+    return render(
+        request,
+        'tasks/previous_week.html',
+        {
+            'first_day': f'{first_day_of_week.strftime('%d/%m/%Y')}',
+            'last_day': f'{last_day_of_week.strftime('%d/%m/%Y')}',
+            'summary': summary,
+            'count_completed_percent': count_completed_percent,
+            'count_completed': count_completed,
+            'count_pending': count_pending
+        }    
+    )
